+++ app/helpers/application_helper.rb
@@ -75,10 +75,6 @@ module ApplicationHelper
     end
   end
 
-  def safe_url?(url)
-    ['http', 'https', 'ftp', 'mailto'].include?(URI.parse(url).scheme) rescue false
-  end
-
   def simple_markdown(text, table_class = '')
     markdown = Redcarpet::Markdown.new( Redcarpet::Render::HTML, no_intra_emphasis: true, fenced_code_blocks: true, autolink: true, strikethrough: true, superscript: true, tables: true, lax_spacing: true, space_after_headers: true, underline: true, highlight: true, footnotes: true )
     result = replace_numbers_with_ascii(text.to_s)