+++ test/helpers/events_helper_test.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-class EventsHelperTest < ActionView::TestCase
-end